# Remove layer of channels